coverage improvements, removed python3-specific assertWarns context manager
